Reorganize: param_ad -> models/, eager/ package, shared boundary -> types/
Three code-organization fixes (no behavior change) for modules misplaced during
the parameter-derivative work:

- neml2/es/param_ad.py -> neml2/models/param_ad.py. The reverse-mode
  parameter-derivative engine has zero equation-system dependencies; it is the
  reverse-mode counterpart to models/chain_rule.py (forward-mode input
  sensitivities) and is consumed by Model + _EagerModel, so it belongs in models/,
  not es/.

- neml2/eager.py -> neml2/eager/ package. __init__.py re-exports _EagerModel from
  eager/_model.py; the import path neml2.eager (which the C++ embedded runtime
  imports) is unchanged.

- neml2/_eager_boundary.py -> neml2/types/_boundary.py. These are the shared
  raw-tensor framework-boundary helpers -- check_tensor / broadcast_to_common_batch
  (AOTI shim + eager), unwrap_outputs (Model + eager), assemble_jvp_outputs /
  assemble_jacobian (eager only) -- not eager-specific. Moving them out of eager/
  removes the backwards dependency where the AOTI shim and Model imported from the
  eager package; they now sit at the raw<->typed boundary in types/.

Imports + docstrings updated across model.py, aoti/_shim.py, the moved files, and
test_eager.py; CLAUDE.md package layout + Rule 1 updated. Verified: test_eager.py
+ test_aoti_shim.py (36), test_aoti_export.py (19), and the C++ test_eager pass.
