+++ neml2/eager/__init__.py
@@ -1,0 +1,40 @@
+# Copyright 2024, UChicago Argonne, LLC
+# All Rights Reserved
+# Software Name: NEML2 -- the New Engineering material Model Library, version 2
+# By: Argonne National Laboratory
+# OPEN SOURCE LICENSE (MIT)
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+"""Embedded-Python eager runtime adapter.
+
+The C++ ``neml2::eager::Model`` (``neml2/csrc/eager/``) embeds a CPython
+interpreter, imports this package, and drives :class:`_EagerModel` to evaluate a
+model straight from its ``.i`` with no compile step. :class:`_EagerModel` (in
+:mod:`._model`) is the raw-tensor / name-keyed ``forward`` / ``jvp`` / ``jacobian``
+/ ``param_jacobian`` / ``param_vjp`` surface the C++ runtime consumes.
+
+The raw-tensor framework-boundary helpers it shares with the AOTI shim (the
+device/dtype check, batch-broadcast, and output-unwrap / assembly; see CLAUDE.md
+Rule 1) are NOT eager-specific and live in :mod:`neml2.types._boundary`.
+"""
+
+from ._model import _EagerModel, _infer_device_dtype
+
+__all__ = ["_EagerModel", "_infer_device_dtype"]
